Phase 1 Inspection
(More notes can be found in comments at the header of BaseStation.java).
Finished the phase 1 inspection save for one or two points (look them
over when you can -- I couldn't make a decision on my own). Also
implemented a buildMessage() function to eliminate a lot redundant code
in each of the BaseStation classes (all the appending values to the ends
of messages and then appending checksums).